refactor currency lookup route
